pyproject.toml
[tool.poetry]
name = "gebunden"
version = "0.1.0"
description = ""
authors = ["David Banks <amoebae@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
gunicorn = "^20.1.0"
pyramid = "^2.0"
waitress = "^2.0.0"
pyramid-jinja2 = "^2.8"
beautifulsoup4 = "^4.10.0"
attrs = "^21.2.0"

[tool.poetry.dev-dependencies]

[tool.poetry.plugins."paste.app_factory"]
main = "gebunden:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
